+++ Classes/TBC/Druid.lua
@@ -799,7 +799,7 @@ function Druid:ApplyScalers(weights, currentSpec)
         end
         
         if (currentSpec:find("FERAL_CAT") or currentSpec:find("Cat")) and w["ITEM_MOD_ARMOR_PENETRATION_RATING_SHORT"] then
-            local arPen = GetCombatRating(25)
+            local arPen = GetCombatRating(25) or 0
             if arPen > 100 then
                 local scaler = 1 + (arPen / 1000)
                 if scaler > 1.4 then scaler = 1.4 end